+++ build.gradle
@@ -15,20 +15,20 @@ repositories {
 
 dependencies {
     compile 'io.javalin:javalin:3.1.0'
-    compile "org.slf4j:slf4j-simple:1.7.26"
-    compile "com.fasterxml.jackson.core:jackson-databind:2.9.8"
+    compile 'org.slf4j:slf4j-simple:1.7.26'
+    compile 'com.fasterxml.jackson.core:jackson-databind:2.9.8'
     compile 'org.xerial:sqlite-jdbc:3.27.2.1'
-    testImplementation('org.junit.jupiter:junit-jupiter:5.4.2')
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.4.2'
 }
 
 test {
     useJUnitPlatform()
     testLogging {
-        events "passed", "skipped", "failed"
+        events 'passed', 'skipped', 'failed'
     }
 }
 
-task stage(dependsOn: "assemble")
+task stage(dependsOn: 'assemble')
 
 application {
     mainClassName = 'com.jhuoose.todoose.Server'
