+++ documentation/TODOOSE.postman_collection.json
@@ -184,7 +184,7 @@
 							"    });",
 							"});",
 							"",
-							"pm.test(\"Status code is 404\", () => {",
+							"pm.test(\"Status code is 404 if item isn’t found\", () => {",
 							"    pm.sendRequest({ url: \"http://localhost:7000/items/51039\", method: \"DELETE\" }, (err, response) => {",
 							"        pm.expect(response).to.have.status(404);",
 							"    });",
@@ -231,6 +231,113 @@
 					"originalRequest": {
 						"method": "DELETE",
 						"header": [],
+						"url": {
+							"raw": "http://localhost:7000/items/7",
+							"protocol": "http",
+							"host": [
+								"localhost"
+							],
+							"port": "7000",
+							"path": [
+								"items",
+								"7"
+							]
+						}
+					},
+					"status": "No Content",
+					"code": 204,
+					"_postman_previewlanguage": null,
+					"header": null,
+					"cookie": [],
+					"body": null
+				}
+			]
+		},
+		{
+			"name": "Edit item",
+			"event": [
+				{
+					"listen": "test",
+					"script": {
+						"id": "1f28be8a-b12f-4544-ab1f-05458edd4a96",
+						"exec": [
+							"pm.test(\"Status code is 204\", () => {",
+							"    pm.response.to.have.status(204);",
+							"});",
+							"",
+							"pm.test(\"Item is edited\", () => {",
+							"    pm.sendRequest(\"http://localhost:7000/items\", (err, response) => {",
+							"        pm.expect(response.json().last().description).to.equal(\"some description\");",
+							"    });",
+							"});",
+							"",
+							"pm.test(\"Status code is 404 if item isn’t found\", () => {",
+							"    pm.sendRequest({ url: \"http://localhost:7000/items/51039\", method: \"PUT\" }, (err, response) => {",
+							"        pm.expect(response).to.have.status(404);",
+							"    });",
+							"});",
+							""
+						],
+						"type": "text/javascript"
+					}
+				},
+				{
+					"listen": "prerequest",
+					"script": {
+						"id": "1f042eff-532a-4075-84f7-e8397e6b5393",
+						"exec": [
+							"pm.sendRequest({ url: \"http://localhost:7000/items\", method: \"POST\" });",
+							"pm.sendRequest(\"http://localhost:7000/items\", (err, response) => {",
+							"    pm.environment.set(\"itemIdentifier\", response.json().last().identifier);",
+							"});",
+							""
+						],
+						"type": "text/javascript"
+					}
+				}
+			],
+			"request": {
+				"method": "PUT",
+				"header": [],
+				"body": {
+					"mode": "formdata",
+					"formdata": [
+						{
+							"key": "description",
+							"value": "some description",
+							"type": "text"
+						}
+					]
+				},
+				"url": {
+					"raw": "http://localhost:7000/items/{{itemIdentifier}}",
+					"protocol": "http",
+					"host": [
+						"localhost"
+					],
+					"port": "7000",
+					"path": [
+						"items",
+						"{{itemIdentifier}}"
+					]
+				}
+			},
+			"response": [
+				{
+					"name": "Successful",
+					"originalRequest": {
+						"method": "PUT",
+						"header": [],
+						"body": {
+							"mode": "formdata",
+							"formdata": [
+								{
+									"key": "description",
+									"value": "whatever",
+									"type": "text"
+								}
+							]
+						},
 						"url": {
 							"raw": "http://localhost:7000/items/7",
 							"protocol": "http",
