+++ src/test/java/com/jhuoose/todoose/models/ItemTests.java
@@ -1,4 +1,4 @@
-package com.jhuoose.todoose;
+package com.jhuoose.todoose.models;
 
 import org.junit.jupiter.api.Test;
